added validation for jwt

+++ middlewares/auth.go
@@ -26,7 +26,7 @@ func IsAuthenticated(ctx *fiber.Ctx) error {
 	if data == nil || err != nil {
 		return ctx.SendStatus(fiber.StatusUnauthorized)
 	}
-	if float64(data.UpdatedAt) > user.Iat {
+	if data.UpdatedAt > user.Iat {
 		return ctx.SendStatus(fiber.StatusUnauthorized)
 	}
 	ctx.Locals("user", data)
